+++ src/industry-solution-packs-composable-applications/index.js
@@ -1,0 +1,3 @@
+export * from './industry-solution-packs-composable-applications-overview-constants.js';
+export { IndustrySolutionPacksComposableApplicationsOverviewProfile } from './contracts/industry-solution-packs-composable-applications-overview-profile.js';
+export { IndustrySolutionPacksComposableApplicationsOverviewDescriptor } from './overview/industry-solution-packs-composable-applications-overview-descriptor.js';